+++ app/Http/Controllers/AdminController.php
@@ -64,12 +64,12 @@ class AdminController extends Controller
             'first_play_video_source_type' => $settings->first_play_video_source_type,
             'first_play_video_name' => $settings->first_play_video_name,
             'first_play_channel_id' => $settings->first_play_channel_id,
-            'mist_server_ip' => $settings->mist_server_ip,
+//            'mist_server_ip' => $settings->mist_server_ip,
             'mist_server_uri' => $settings->mist_server_uri,
-            'mist_server_api_url' => $settings->mist_server_api_url,
-            'mist_server_username' => $settings->mist_server_username,
-            'mist_server_password' => $decryptedPassword ?? null,
-            'mist_access_control_secret' => $settings->mist_access_control_secret,
+//            'mist_server_api_url' => $settings->mist_server_api_url,
+//            'mist_server_username' => $settings->mist_server_username,
+//            'mist_server_password' => $decryptedPassword ?? null,
+//            'mist_access_control_secret' => $settings->mist_access_control_secret,
         ]);
     }
 
@@ -83,12 +83,12 @@ class AdminController extends Controller
             'first_play_video_source_type' => 'nullable|string',
             'first_play_video_name' => 'nullable|string',
             'first_play_channel_id' => 'nullable|integer',
-            'mist_server_ip' => 'nullable|string',
+//            'mist_server_ip' => 'nullable|string',
             'mist_server_uri' => 'nullable|url',
-            'mist_server_api_url' => 'nullable|url',
-            'mist_server_username' => 'nullable|string',
-            'mist_server_password' => 'nullable|string',
-            'mist_access_control_secret' => 'nullable|string',
+//            'mist_server_api_url' => 'nullable|url',
+//            'mist_server_username' => 'nullable|string',
+//            'mist_server_password' => 'nullable|string',
+//            'mist_access_control_secret' => 'nullable|string',
         ]);
 
         // Encrypting the Mist Server Password
@@ -104,12 +104,12 @@ class AdminController extends Controller
         $settings->first_play_video_source_type = $request->first_play_video_source_type;
         $settings->first_play_video_name = $request->first_play_video_name;
         $settings->first_play_channel_id = $request->first_play_channel_id;
-        $settings->mist_server_ip = $request->mist_server_ip;
+//        $settings->mist_server_ip = $request->mist_server_ip;
         $settings->mist_server_uri = rtrim($request->mist_server_uri, '/') . '/';
-        $settings->mist_server_api_url = $request->mist_server_api_url;
-        $settings->mist_server_username = $request->mist_server_username;
-        $settings->mist_server_password = $encryptedPassword ?? null;
-        $settings->mist_access_control_secret = $request->mist_access_control_secret;
+//        $settings->mist_server_api_url = $request->mist_server_api_url;
+//        $settings->mist_server_username = $request->mist_server_username;
+//        $settings->mist_server_password = $encryptedPassword ?? null;
+//        $settings->mist_access_control_secret = $request->mist_access_control_secret;
 
         $settings->save();
 
@@ -117,6 +117,7 @@ class AdminController extends Controller
       if ($request->has(['first_play_video_source', 'first_play_video_source_type'])) {
 
         $dataToUpdate = $request->only([
+            'mist_server_uri',
             'first_play_video_source',
             'first_play_video_source_type',
             'first_play_video_name',